fix(db): V23 fractional volume — crypto volume is not an integer
V2 defined market_data.volume as BIGINT (stock-market oriented). The
Binance Vision seed failed on the very first row with "invalid input
syntax for type bigint: 3617.988" — crypto volume is fractional.
Migrated to NUMERIC(20, 8) matching the precision used in the derivatives
tables. Java entity + DTO + three callers switch from Long to BigDecimal.
Smoke seed (BTCUSDT 15m 2024-01) now lands 2976/2976 rows.

+++ QuantPlatformApplication/src/main/java/com/QuantPlatformApplication/QuantPlatformApplication/service/DataSeederService.java
@@ -94,7 +94,7 @@ public class DataSeederService implements CommandLineRunner {
             entity.setHigh(BigDecimal.valueOf(round(high)));
             entity.setLow(BigDecimal.valueOf(round(low)));
             entity.setClose(BigDecimal.valueOf(round(close)));
-            entity.setVolume(volume);
+            entity.setVolume(BigDecimal.valueOf(volume));
 
             batch.add(entity);
             price = close;
